Habits tracking system
I have built this script to make an API call to take my habits data, upload it into a database (SSMS) and generate a pdf that is then emailed to me daily (last thing in the day). It uses a credentials JSON file (for all API keys, passwords, etc.) and task scheduler to run the program daily (regardless of whether the computer is on or off).